Reduce memory allocations in render sort, replace Math.random() with custom RNG

+++ src/main/core/content/entity/BlockEntity.java
@@ -11,7 +11,7 @@ import java.io.IOException;
 
 public interface BlockEntity extends BlockDrawComponent {
 
-    void setPosition(short x, short y);
+    void setPosition(int x, int y);
 
     short x();
     short y();
